+++ src/wingsTranslator/src/wingsTranslatorThread.cpp
@@ -284,14 +284,14 @@ bool wingsTranslatorThread::threadInit() {
     Bottle info;
     igaze->getInfo(info);
     printf("getting info from iKinGazeCtrl \n");    
-    int head_version = info.check("head_version", Value(1)).asInt();
+    double head_version = info.check("head_version", Value(1.0)).asDouble();
     
-    printf("head_version extracted from gazeArbiter %d \n",head_version);
+    printf("head_version extracted from gazeArbiter %g \n",head_version);
     ikl = new iCubEye();
     ikr = new iCubEye();
     
-    head_version = 2;
-    if(head_version == 1) {
+    head_version = 2.0;
+    if(head_version == 1.0) {
         eyeL = new iCubEye("left");
         eyeR = new iCubEye("right");
     }
